sesion10/integracion/backend-punto-de-venta/requirements.txt
django
djangorestframework
djangorestframework_simplejwt

django-cors-headers